animation_studio/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.23.2
openai==1.77.0
aiohttp==3.9.1
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pillow==11.2.1
requests==2.31.0
opencv-python==4.10.0.84
httpx==0.25.2 